Update build.gradle
Build.gradle had a commented out section for compose dependencies.
Apparently this caused gradle/studio to bump the version of
androidx.activity to 1.2.0-alpha02, evan though it was set as 1.1.0 in
the file. I was unware of this so when I removed the commented out
compose 'implementation' lines, gradle/studio used the version of
activity set in the file and the build failed. This was because I was
using the ActivityResult API and it was only available starting from the
1.2.0 alphas. I also had to bump fragment to 1.3.0-alpha07.